Ultralib (#10)
* git subrepo clone git@github.com:decompals/ultralib.git lib/ultralib/

subrepo:
  subdir:   "lib/ultralib"
  merged:   "b376559"
upstream:
  origin:   "git@github.com:decompals/ultralib.git"
  branch:   "main"
  commit:   "b376559"
git-subrepo:
  version:  "0.4.6"
  origin:   "git@github.com:ingydotnet/git-subrepo.git"
  commit:   "110b9eb"

* fix

* flib libultra syms

+++ src/main/loadfragment2.c
@@ -16,7 +16,7 @@ void DoRelocation(void* allocatedVRamAddr, OverlayRelocationSection* ovl, uintpt
     uintptr_t relocatedAddress;
     u32 i;
     u32* luiInstRef;
-    uintptr_t allocu32 = (uintptr_t)allocatedVRamAddr;
+    u32 allocu32 = (uintptr_t)allocatedVRamAddr;
     u32* regValP;
     u32* luiRefs[32];
     u32 luiVals[32];
